Item14033: Convert JsonRpcContrib to Request::JSON
This is incomplete.  It appears functional but there are several issues:

Request::param() and Request::JSON::param() collide.  They can coexist
as long as the JSON version is read-only on the JSON parm data. so that
write requests can be passed to the CGI implementation.  If it needs to
be read-write, this will probably have to be split into a separate
jsonparam() routine.

Errors encountered in Request::JSON are not passed through to the UI.
this needs some additional work.