Item13897: First working test in unit test!
Yes, QueryTests does it! Don't ask why is this particular one – it was just
small enough start with... Yet, not sure if successfull pass isn't caused
by a hidden bug.

- Added Unit::TestApp class to avoid polluting Foswiki::App with test-only
related code.

- Added Foswiki::Object::Parent and *::Child classes. They were spawned as
a side-effect of a nonviable idea and are not used; but as they were
planned for some future development I decided have them included now.

- Foswiki::Aux::Localize can behaviour can now be tuned by additional
flags. The flags are set using setLocalizeFlags() method provided by a
class with Localize role. Currently the only flag supported is
clearAttributes which indicates whether attributes of the object being
localized must be cleared or left alone. Used by FoswikiTestCase pushApp
method.

- New Foswiki::Object feature – object cloning. As any cloning it has it's
problems and not every clone is the exact copy of the original – yet it
tries to do its best. Especially when it's going about rather simple
classes like Foswiki::Config. An object may provide additional aid to the
process by declaring _clone_attrName() methods where attrName is the
attribute name to be cloned. These methods must return a cloned copy of
attrName attribute.

- Foswiki::AppObject now declares the app attribute as protected.

- Added Foswiki::AppObject _clone_app() method to support object cloning.

- Added new concept to FoswikiTestCase – push/pop of application object.
It's developed as a replacement for the old style code when a request
object is preserved -> new temporary $session is created -> some code
executed -> new semi-permanent $session is created and the original request
object is restored. For now we just pushApp -> do some code -> popApp.

Note that createNewFoswikiApp() and popApp() call _fixupAppObject() to set
all attributes of the current test case object which are descendants of
Foswiki::App to have their app attribute point to the currently active
application. Generally it shall be transparent to the end user. But if in
some case a new application object is created and set active by a test case
it must call _fixupAppObject() too.