Item13897: PlackTestCase is finished.
This is another and hopefully the last megacommit to this branch. It may
have impact on overall code stability of this branch. I'm sorry for that
but to write those three words in the commit subject took a bit more than
just few changes. Next time a branch will be used.

With this commit I consider PlackTestCase ready for use. It has all the
basic functionality needed to test the application with Plack::Test and
the same time benefit from the old UnitTestContrib code.

No documentation yet but it come in a day or two.

- Added PlackPostTests as sample test suite for PlackTestCase.

- Extracted all code which can serve for both
FoswikiTestCase/FoswikiFnTestCase and PlackTestCase into a new role
FoswikiTestRole. This also includes splitting of test initialization code
into several `setupSomething()' methods.

- Renamed pushApp/popApp test case methods to saveState/restoreState to
correspond better with their purpose.

- Step-by-step replacement of Foswiki::Func::apiFunc() to
$Foswiki::app->apiFunc() or $this->app->apiFunc() wherever possible.

- Step-by-step switching from 'use Moo' to 'use Foswiki::Class'.

- Dropping support for old-style positional parameters for object
constructors.

- Added new exception Foswiki::Exception::FileOp for failed file
operations. It automates use of $!.

- Unit::TestApp now registers callbacks as early as possible – before
Foswiki::App::cfg attribute get initialized. This is to make it possible
for tests to manipulate with LSC before application is completely
initialized.

- Continue replacing $Foswiki::cfg with app->cfg->data.

- Added isGuest() method to Foswiki::Users.

- Added parameter `app' to Foswiki::Class. It automates applying of
Foswiki::AppObject role to a class.

- Fixed a strange bug where Foswiki::App::user attribute read was
overoptimized for reading when passed as a method argument. The
optimizations resulted in stack corruption if the attribute has been
changed before completing the method. So, it is now mandatory for the
attribute to be lazy and have a builder.

- `postConfig' callback is now raised by Foswiki::App constructor right
after checking the configuration attribute status and before any other
initialization is done.

- Fixed exception handling in Foswiki::UI::Register.

- Foswiki::Aux::Localize::setLocalizeFlags is now returning key/value pairs
list instead of hashref. This is to simplify it's use with inheritance.

- Fixed test case registerUser() method failing to detect errors generated
by Foswiki::UI::Register.