Item13897: Documentation for PlackTestCase.
- Added support for non-Foswiki prefixed modules to INCLUDE's doc handler.
It is now handles all modules located in subdirectories for the directory
where Foswiki.pm is located as local documentation. So, for example, if a document
references Unit::TestCase as a plain text doc handler will convert it into
a valid link to Unit/TestCase.pm as it was doing this with modules under
Foswiki:: hieararchy.

- Added support for external modules to INCLUDE's doc handler. If a module
is not located in Foswiki directory tree as described above it is
considered external and a link to CPAN is generated.

- Added some documentation to Unit::FoswikiTestRole.

- Unit::FoswikiTestRole::setupDirs doesn't fail if directory already
exists.

- Unit::FoswikiTestRole::setupDirs throws Foswiki::Exception::FileOp if
mkdir fails.

- Unit::PlackTestCase::initRequest() initializes server application
instance similar to initTest() method. This was a forgotten piece of code.