Item13897: Experimental rewrite of Foswiki::IncludeHandlers::doc
It was a hard to extend mess. Reimplemented using simple lexical parsing
based on Perl's regexps. It is now much easier to adapt it to the new
OO terminology and constructs.

The old behaviour can be returned back by setting
Foswiki::IncludeHandlers::doc::USER_LEXICAL_PARSER constant to 0. Could
also be controlled by some LSC value but I see no point in this.

- Slightly improved error handling by INCLUDE macro.