Item14205: Fix some issues with FCGI, undefined, and Certs
Only the path or file can be passed for the CA Certificate store. Some
verions will reject the connection if both methods are provided.  So if
both exist, us the bundle file if it's available.

The STDERR capture doesn't work on FCGI,  so bypass capturing on that
platform.

Correct some undefined errors.