Item14219: Change from %BUTTON to a template
Still not part of a MANIFEST or the release. Needs some further review.

Note that this has made obvious to me anyway that the USERSTYLEURL
settings are pretty poor.   You can only have one.  You have to choose
from one of the Cookbook examples, or merge stylesheets externally.

Also PatternSkin doesn't make any use of ADDTOZONE.  Probably ought to
be looked at as a 2.2 feature.