Item14205: More compatibility issues with IO::Socket::SSL
Versions older than 1.973 don't support setting of both SSLCaPath and
SSLCaFile.  They are mutually exclusive.   Newer versions support
setting both.

Ubuntu 14.04 ships with 1.965, so the autoconfig fails on that platform.

Check the versions, and ensure that only one or the other are set with
older versions of IO::Socket::SSL