Item14181: Improve bootstrap operation behind proxy servers
This attempts to do a better job of detecting operation behind a proxy.

Detecting a proxy is relatively straight forward, by examining
X_FORWARDED_HOST.

Detecting HTTPS is a bit more iffy.  The browsers *appear* to set
HTTP_UPGRADE_INSECURE_REQUESTS, however it is not set for the jsonrpc
requests. These however have the referer set to the X_FORWARDED_HOST,
and that can be used to detect https.

This code was only tested with a single stage proxy.  It was tested with
Firefox and Chrome.