Item14181: Another try at SSL behind a proxy
Turns out the browsers always set HTTP_UPGRADE_INSECURE_REQUESTS,
so that can't be used.  Last resort,  set a query param of
SSL=1, to give bootstrap a hint that it needs to be using SSL URLs.

I can't find any other way to handle this. The proxy completely hides
that it is using SSL to the client.