Item14237: Added 'perl' and 'legacy' spec formats.
'legacy' is for the Foswiki 2.x style .spec format. Currently parsed but
not completely supported. Compatibility is not guarantted though the
parser was trying to resemble the LoadSpec.pm code as much as possible.

'perl' is for perl code format where it'd be possible for spec writer to
write any Perl code which would then be inserted into a sub body and
called. Sub return value is supposed to be a list of spec items.

- Made Foswiki::Config::_getSubHash a public method.