Item14315: Don't load latest rev in attachments table
The attachments table uses QUERY to get the attachments count.  For this
to be correct, it needs to specify the rev of the topic being used.

Also the QUERY macro, when asked to expand {Configuration}{Keys} does
not need to load a topic rev.

Last, when QUERY does have to load the latest revision of a topic, load
a  new topic object, don't try to reload the passed object.  That's what
causes the ASSERT - loading a new rev into an existing object.