Item14237: New method expandStr() to replace expandValue()
expandStr() brings in support for ${Key.SubKey} macro format as a part
of movement to get rid of %Foswiki::cfg global. Another form of the
macro is ${Key}{SubKey}.

expandStr() uses parametrized calling convention and supports expanding
both free-form strings and keys defined by their names. Handling of
undefs is similar to that in expandValue: i.e. undef could be replaced
with a string (though expansStr can replace it with any user-defined
value), or undef would be returned for the expansion, or an exception
could be thrown.