Item14237: Write/Read new style LSC
As a measure to make LSC the only source of configuration data for
Foswiki methods writeLSC/readLSC were introduced. They're currently
handling a new LSC format which is not Perl code. But this is not
important as any format could be chosen. What is important is that these
methods deal with writing/reading the low-level config data only. Both
are pluggables to simplify creation of plugins handling special cases of
LSC like use of different backends including remote or database
storages.

- Renamed 'data' parameter to 'dataObj' wherever appropriate; i.e. if
the parameter is a DataHash instance.

- specsMode method can now generate standalone tied data hashes without
touching config's object data attribute.