Item14237: Switched over to the new spec-based write/read LSC
- Implemented staged app initializing. Stages are named and can be
determined by examining app's initStage attribute (because contexts are
not ready to be used yet).  The first stage, 'initConfig', is when the
config is first read. The second stage, 'extLoad', is when extensions
(not plugins!) are being loaded. The third stage, 'reConfig', is when
config is being read second time. This is when any extension could take
control over config loading. And finally the execution flow reaches
'postConfig' stage when callback with the same name is being called and
when context become useful enough to overtake the attribute
functionality.

- readConfig is not used anymore. It's code is still part of the
Foswiki::Config but mainly for occasional testing only. It is replaced
with pluggable method read().

- Added expandAll method.

- Replaced \ (backslash) based escaping when expanding macros with
${<char} scheme. Like, ${$} is expanded into just $, ${\} – into \. ${}}
– into }.