Item14415: Use NFKD comparisons when inserting entries.
This leaves in some commented debug prints.   There is still an issue
that can cause out-of-order inserts.  Any entry that is not a wikiword
"terminates" the list, and all inserts go before the non-wikiword entry.
This really should not happen unless the topic was manually maintained,
but if it happens, future registrations end up in the wrong place.