Item14493: Restructure ICON Templates
Move icon.tmpl from FamFamFamContrib to core templates.  It's needed by
the core ICON.pm,  and should not be part of the FamFamFamContrib.

Remove the icon:image template from foswiki.tmpl, so the primary
definition is directly in foswiki.tmpl.  The original template has a mix
of single and double quotes which breaks the Render::IconImage quote control.

Eliminate the icon:image_singlequote template.  It isn't used, and the
icon:image template already uses single quotes.