Item14380: A couple of additional tweaks:
This version has been tested with an apache reverse proxy in front of a
plain CGI foswiki and a mod_perl foswiki. Tested with http, https, and
non-standard ports.  All seem to be working fine.

Note that FuncTests::test_getUrl fails, but it appears that the test is
broken.

Also HomePagePlugin tests fail, but it does work when used live. So
more possibly broken tests.

FuncTests has 1 unexpected results (of 58):
   * F: FuncTests::test_getUrlHost
HomePagePluginSuite has 2 unexpected results (of 6):
   * F: HomePagePluginSuite::test_login
   * F: HomePagePluginSuite::test_view