Item14532: Change parameter name abbreviation
GetOptions is not case sensitive, so -N for name was colliding with -n
for process count.   Changing to -a- for nAme.  Also set a default
name if parameter is not included.

Added an env variable to make the process name easier to override.
Added it to the systemd foswiki.service file as well.