Item14563: Item10219: Fix AUTOLINK issues.
This Isn't a full fix, but it gets closer.

Refactor the AUTOLINK code into META from the UI, and tighten up the
window of opportunity to collide on the autolink topic.

Really the atomic lock should be for the SomeTopicAUTOLINK prefix rather
than the individual topic.  This would eliminate any race conditions,
but the Store API cannot override the topic name for the atomic lock
API.