Item14602:  Implement lower case topic name support.
This also fixes a bug in the WebTopicCreator javascript, where it
was failing to auto-capitalize non wikiword names when wikiwords were
requested.

This also eliminates an old holdover from TWiki in the Rename script. If
a topic did not exist as entered, but did exist with the first character
lower-case, the rename would allow the hidden lower-case topic name to
be renamed.

It changes sort of topic names returned from SEARCH to be case
insensitive.